requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
neo4j==5.28.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
urllib3==2.5.0
